bootdata: Change global memory address to 64bit

+++ hw/ip/test/src/tb_lib.hh
@@ -123,8 +123,8 @@ struct BootData {
     uint32_t tcdm_start;
     uint32_t tcdm_size;
     uint32_t tcdm_offset;
-    uint32_t global_mem_start;
-    uint32_t global_mem_end;
+    uint64_t global_mem_start;
+    uint64_t global_mem_end;
     uint32_t cluster_count;
     uint32_t s1_quadrant_count;
 };
